Rename perf.html, make it the home page for the table preview
