--- packages/table/preview/perf.html
@@ -1,136 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Table Perf</title>
-  <link rel="stylesheet" type="text/css" href="dist/perf.css">
-  <style>
-    body {
-      position: absolute;
-      top: 0;
-      right: 0;
-      left: 0;
-      bottom: 0;
-
-      margin: 0;
-      padding: 0;
-
-      background-color: #F5F8FA;
-    }
-
-    body.pt-dark {
-      background-color: #293742;
-    }
-
-    #page-content {
-      position: absolute;
-      top: 50px;
-      left: 0;
-      bottom: 0;
-      right: 0;
-    }
-
-    .container {
-      display: flex;
-      overflow: hidden;
-      height: 100%;
-    }
-
-    .table {
-      flex: 1;
-      order: 2;
-      z-index: 2;
-    }
-
-    .table.is-inline {
-      margin-left: -12.5%;
-      max-height: 50%;
-      max-width: 50%;
-      transform: translateX(50%) translateY(50%);
-    }
-
-    .sidebar {
-      background-color: #EBF1F5;
-      flex-basis: 300px;
-      height: 100%;
-      order: 1;
-      overflow-y: auto;
-      padding: 0 15px 15px;
-      z-index: 3;
-    }
-
-    .pt-dark .sidebar {
-      background-color: #202B33;
-      box-shadow: 0 0 0 1px #182026; /* override pt-elevation-0 styles */
-      z-index: 1;
-    }
-
-    .sidebar h4 {
-      border-top: 1px solid #BFCCD6;
-      font-size: 14px;
-      margin: 15px -15px 0;
-      padding: 20px 15px 12px;
-      text-transform: uppercase;
-    }
-
-    .sidebar h4:first-child {
-      border-top: none;
-      margin-top: 0;
-    }
-
-    .sidebar h6 {
-      color: #5C7080;
-      font-size: 14px;
-      font-weight: normal;
-      line-height: 12px;
-      padding: 10px 0 3px;
-    }
-
-    .pt-dark .sidebar h4 {
-      border-top-color:#182026;
-      color: #A7B6C2;
-    }
-
-    .pt-dark .sidebar h6 {
-      color: #A7B6C2;
-    }
-
-    .sidebar label:last-child {
-      margin-bottom: 0;
-    }
-
-    label.tbl-select-label {
-      margin-top: -3px;
-      margin-bottom: 7px;
-    }
-
-    .tbl-select-label .pt-select {
-      float: right;
-    }
-
-    .tbl-zebra-stripe {
-      background: #F5F8FA;
-    }
-
-    .tbl-styled-region-success {
-      border: 1px solid #0F9960;
-      background-color: rgba(15, 153, 96, 0.1);
-    }
-
-    .tbl-styled-region-warning {
-      border: 1px solid #D9822B;
-      background-color: rgba(217, 130, 43, 0.1);
-    }
-
-    .tbl-styled-region-danger {
-      border: 1px solid #DB3737;
-      background-color: rgba(219, 55, 55, 0.1);
-    }
-  </style>
-</head>
-<body>
-  <div id="nav"></div>
-  <div id="page-content"></div>
-  <script src="dist/perf.bundle.js"></script>
-</body>
-</html>
